remove projects references, fixes issue #458
